xwayland-shm: fortify fallocate against EINTR
If posix_fallocate or ftruncate is interrupted by signal
while working, we return -1 as fd and the allocation process
returns BadAlloc error. That causes xwayland clients to abort
with 'BadAlloc (insufficient resources for operation)'
even when there's a lot of resources available.

Fix it by trying again when we get EINTR.

Signed-off-by: Marek Chalupa <mchqwerty@gmail.com>